Added support for half stars (or quarters etc..)

+++ generators/ajaxful_rating/templates/migration.rb
@@ -3,7 +3,7 @@ class CreateRates < ActiveRecord::Migration
     create_table :rates do |t|
       t.belongs_to :rater
       t.belongs_to :rateable, :polymorphic => true
-      t.integer :stars, :null => false
+      t.float :stars, :null => false
       t.string :dimension
       t.timestamps
     end
